+++ drift/test/integration_tests/database_test.dart
@@ -4,6 +4,7 @@ library;
 
 import 'package:drift/drift.dart';
 import 'package:drift/native.dart';
+import 'package:drift/src/runtime/cancellation_zone.dart';
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
 
@@ -32,12 +33,10 @@ void main() {
         return driftDb.select(driftDb.categories).get();
       }),
       throwsA(
-        isA<CouldNotRollBackException>().having(
-          (e) => e.cause,
-          'cause',
-          isA<SqliteException>().having((e) => e.causingStatement,
-              'causingStatement', 'BEGIN TRANSACTION'),
-        ),
+        isA<SqliteException>()
+            .having((e) => e.causingStatement, 'causingStatement',
+                'BEGIN TRANSACTION')
+            .having((e) => e.extendedResultCode, 'resultCode', 262),
       ),
     );
 
@@ -152,5 +151,21 @@ void main() {
 
     final results = await Future.wait<void>([fut1, fut2]);
     expect(results, [7, 5]);
+  });
+
+  test('can cancel opening transactions', () async {
+    final db = TodoDb(NativeDatabase.memory());
+    addTearDown(db.close);
+
+    final token = CancellationToken<void>()..cancel();
+
+    runCancellable(() async {
+      await db.transaction(() async {
+        throw 'should not be reached';
+      });
+    }, token: token);
+    expect(token.result, throwsA(isA<CancellationException>()));
+
+    await db.customSelect('SELECT 1').get();
   });
 }